nambah select search, datatables, rapihin halaman print, rapihin procedure detail produk, hapus field supplier di table produk, dan laen-laen

+++ app/Http/Controllers/SuperAdminController.php
@@ -11,9 +11,7 @@ use Illuminate\Support\Facades\Hash;
 class SuperAdminController extends Controller
 {
     public function index(Request $request)
-
     {
-        // $user = user::all();
         $search = $request->search;
         $level_user = DB::table('level_user')->select()->get();
         $user = DB::table('user')
@@ -30,6 +28,12 @@ class SuperAdminController extends Controller
     // tambah
     public function store(Request $request)
     {
+        request()->validate(
+            [
+                'foto' => 'image|mimes:jpg,png,jpeg,gif,svg|max:1000'
+            ]
+        );
+
         $user = new User;
         // $id_user = substr(md5(rand(0, 99999)), -4);
         $id_user = collect(DB::select('SELECT new_iduser() AS new_iduser'))->first()->new_iduser;
@@ -61,6 +65,12 @@ class SuperAdminController extends Controller
     // update
     public function update(Request $request)
     {
+        request()->validate(
+            [
+                'foto' => 'image|mimes:jpg,png,jpeg,gif,svg|max:1000'
+            ]
+        );
+
         $user_id = $request->input('user_id');
         $user = User::find($user_id);
         $user->nama = $request->input('nama');
@@ -85,21 +95,19 @@ class SuperAdminController extends Controller
     // hapus
     public function destroy(Request $request)
     {
-        $user_id = $request->input('delete_user_id');
-        $user = User::find($user_id);
-        $destination = 'storage/post-images/' . $user->foto;
-        if (file::exists($destination)) {
-            file::delete($destination);
-        }
-        $user->delete();
-        return redirect()->back()->with('success', "Data Berhasil di Hapus");
-    }
 
-    // search
-    // public function search(Request $request)
-    // {
-    //     $get_name = $request->search;
-    //     $user = user::where('nama', 'LIKE', '%' . $get_name, '%')->get();
-    //     return view('SuperAdmin.index', compact('user'));
-    // }
+        $user_id = $request->input('delete_user_id');
+        if ($user_id == 'USR01') {
+            return redirect()->back()->with('warning', 'Super Admin tidak dapat dihapus!');
+        } else {
+            # code...
+            $user = User::find($user_id);
+            $destination = 'storage/post-images/' . $user->foto;
+            if (file::exists($destination)) {
+                file::delete($destination);
+            }
+            $user->delete();
+            return redirect()->back()->with('success', "Data Berhasil di Hapus");
+        }
+    }
 }
